+++ CHANGELOG.md
@@ -6,11 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/ja/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unreleased - 2025-09-09
+
+## [2.0.0] - 2025-09-09
+
 ### Added
 - High Capacity Jetpack
 
 ### Removed
 - Mechanical Boots
+
+[2.0.0]: https://github.com/pluslatte/gt6gears/releases/tag/v2.0.0
 
 ## [1.1.1] - 2025-05-17